Update provider to virtualbox
Docker does not currently work with vagrant-proxyconf plugin so the
proxies are not set on the container. The remote ansible user gets provided
the proxies so can run apt-get and such commands successfully.
But when lauching sudo apt-get update from within the /opt/refstack/setup_env
script that user does not have the proxies, leading to failures.

So we are using virtualbox instead, which is more inline with the jenkins
tests that get run against the deployments.